Yannis: SingletonArrayToValueTransformer works for Scala as well

+++ generator-out/run_scala.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 END=22
 START=1
+
+if [ $# -ne 2 ]; then
+    echo "Invalid number of command line arguments."
+    echo "USAGE: ./run_scala.sh <DATA_FOLDER> <SF>"
+    exit
+fi
+
 rm -rf bin
 mkdir bin
 CPATH=$HOME/.ivy2/local/lego-core/lego-core_2.11/0.1-SNAPSHOT/jars/lego-core_2.11.jar:$HOME/.ivy2/local/ch.epfl.data/pardis-library_2.11/0.1-SNAPSHOT/jars/pardis-library_2.11.jar:$HOME/.ivy2/local/ch.epfl.data/pardis-core_2.11/0.1-SNAPSHOT/jars/pardis-core_2.11.jar
